Fix tipo de dato para Price en Retrofit

+++ Correapp/app/src/main/java/com/fiuba/tdpii/correapp/models/web/TripSerialized.java
@@ -42,7 +42,7 @@ public class TripSerialized {
 
             @Expose
             @SerializedName("_price")
-            public Long price;
+            public Double price;
 
     }
 }
